OptiPrime
First round of optimizations. When I created a variable for the mod, my
favorite non-prime odd number (9) and his annoying friends appeared. I
will mess around with this some more tomorrow for sure and will fix it
(I hope!). Runtime is 35 seconds. I'm going to kickoff the non-optimized
version now and will see how horrifyingly slow it was in the morning. :)